--- a/Makefile
+++ b/Makefile
@@ -5,3 +5,6 @@ tests:
 
 clean:
     find . -name "*.pyc" | xargs -I {} rm -v "{}"
+
+build:
+	python setup.py build_ext -i -b .