--- a/Makefile
+++ b/Makefile
@@ -1,10 +1,13 @@
-.PHONY: tests clean
+.PHONY: build tests clean deploy
+
+build:
+	python setup.py build_ext -i -b .
 
 tests:
     nosetests --logging-level=INFO
 
 clean:
-    find . -name "*.pyc" | xargs -I {} rm -v "{}"
+	find . -name "*.pyc" | xargs -I {} rm -v "{}"
 
-build:
-	python setup.py build_ext -i -b .
+deploy:
+	python setup.py sdist bdist_wheel upload