… .travis.yml
… notebooks/develop/Low_pass_filter.ipynb
… notebooks/develop/Test_global_em1d_inversion_booky_example.ipynb
… notebooks/examples/EM1D_inversion_FD_height_correction.ipynb
… notebooks/examples/EM1D_inversion_FD_height_correction_halfspace.ipynb
… notebooks/examples/EM1D_inversion_TD_layers.ipynb
… notebooks/examples/EM1D_inversion_TD_layers_dual_moment.ipynb
… notebooks/examples/Test_global_em1d_inversion.ipynb
… notebooks/examples/Test_global_em1d_inversion_td.ipynb
… simpegEM1D/Fortran/m_rTE_Fortran.f90
… simpegEM1D/GlobalEM1D.py
… simpegEM1D/Waveforms.py
… tests/testGlobalEM1D_TD_jac_layers.py