Reviewing pull request: #19 

From @leonfoks 

I've updated the 1D forward modeller to use numpy broadcasting and Fortran for a nice speed up (see below for timings)

For timings I used the default 19 layer model from the test example scripts.

In the setup.py file i've added the Fortran extension.  On my Mac numpy.distutils would not add the extra compile flags needed from the extra_f90_compile_flags argument, so i've stuck them in the link args.  This shouldn't be a problem.

## Timing

### Forward Modelling
EM1D - Survey
Two major bottlenecks
 1) projectFields - piecewise pulse - piecewise ramp - Scipy.fixed_quad
 2) forward - hz_kernel_circular_loop - rTEfunfwd/RTEfun_vec

#### Solutions
1) Had a double loop in python in piecewise_ramp and lots of redundantly recalculated arrays inside the loop as a pythonic inline list expansion.
By first expanding memory in C and utilizing broadcasting, the call time for only projectFields() for 1000 forward models was reduced from 20.6 s to 1.27s.  This could be made faster by reducing to Fortran, I would just need to write the 1D interpolator in Fortran. Not necessary just yet.

2) Because of the recursion relation, there's not much you can do other than use numpy broadcasting inside a python loop...Which is what was already done in RTEfun_vec.
However, this can be sped up by reducing to Fortran (or C), I chose the former.
After I did this, the call time for the forward() function for 1000 forward models went from 24.6s to about 10-11s.  I was not expecting great speed up for this one function given that the only slow down was the python loop. Also I used a loop so no need for all the extra memory required when using numpy.

A single forward model is about 16ms instead of ~45ms. Which is a lot of time for our parallel code once you factor in the number of iterations and the number of data.

### Sensitivity i.e. getJ_sigma()
1) ProjectFields - same as above.
2) forward - hz_kernel_circular_loop - rTEfunjac
Quite a few issues here with redundant, repeated code, if statements inside loops, and multiple appends to pythonic lists.

#### Solutions
1) Used the same functions as above, instead of piecewisePulse i wrote piecewise_pulse_fast().
Since there are two calls to piecewise_ramp the new fast function reduced the time from 46.3s to 1.76s for 100 sensitivity calculations
2) Recoded any matmul operations when I could take advantage of symmetry and negation of the components to minimize flops
    Used loops in Fortran to remove memory overhead that broadcasting requires
    100 calls to forward() went from 16.8s to 6.66s
    
## Some general comments on speed
1) Avoid inline loops and list building i.e. 
np.array([fixed_quad(step_func, t, t+t0, n=n)[0] for t in time])
It's extremely slow compared to numpy arrays.

2) Don't create empty arrays assigned to a variable and then immediately reassign that variable. e.g.
rTE = np.empty((n_frequency, n_layer), dtype=complex)
rTE = rTEfunfwd(
    n_layer, f, lamda, sig, chi, depth, self.survey.half_switch
)
rTE gets reassigned to the output of rTEfunfwd, so the np.empty statement just allocated space for no reason.  The overhead is minor, but it's unnecessary. If you want to assign the values to the memory of the first empty call, use rTE[:, :] = function(...).