@sgkang 2018-10-02
Hi @leonfoks, am I supposed to `python setup.py install` to compile m_rTE_Fortran.f90?
It seems not working... 

I used to use 
```python
from distutils.core import setup
```

but it seems you are using 

```python 
from numpy.distutils.core import setup
```

Can you explain bit more what am I supposed to make this work on my machine?