@leonfoks 2018-10-02
@sgkang I first navigate to the simpegEM1D root folder and use "pip install ." to install the package.

unfortunately the distutils.core setup cannot handle Fortran extension.  The numpy version calls f2py on the Fortran code and then compiles the shared library. f2py comes packaged with Numpy so no need to install anything.