@sgkang 2018-10-09
@leonfoks No worries... it is working both of my mac and linux machine. 
Not sure why its failing in travis, but will figure it out.