@sgkang 2018-10-22
@leonfoks, I am thinking about merging this branch to master today, let me know if you any comment or suggestion that you want to make!