@sgkang 2018-10-24
Thanks a lot particularly for @leonfoks, this improvement is huge!
Thank you @prisae for reviews and constructive comments. I learned a lot with this pull request from you guys.